requirements.txt
asgiref==3.5.0
boto3==1.17.22
botocore==1.20.22
certifi==2021.10.8
charset-normalizer==2.0.12
cloudinary==1.29.0
dj-database-url==0.5.0
Django==4.0.3
django-ckeditor==6.2.0
django-cloudinary-storage==0.3.0
django-environ==0.8.1
django-js-asset==1.2.2
django-storages==1.11.1
gunicorn==20.0.4
idna==3.3
jmespath==0.10.0
Pillow==9.0.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.6.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.27.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
tzdata==2022.1
urllib3==1.26.9
whitenoise==5.2.0
